# openapi.yaml
openapi: 3.0.3
info:
  title: API FlopBox
  description: Plate-forme proxy FlopBox pour centraliser la gestion de fichiers distants stockés dans des serveurs FTP tiers.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Serveur local

# Définition des paramètres réutilisables (pour ne pas répéter le code)
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    serverId:
      name: ID
      in: path
      required: true
      description: Identifiant unique du serveur
      schema:
        type: integer
    ftpUsername:
      name: X-FTP-Username
      in: header
      required: true
      description: Nom d'utilisateur pour la connexion FTP distante
      schema:
        type: string
    ftpPassword:
      name: X-FTP-Password
      in: header
      required: true
      description: Mot de passe pour la connexion FTP distante
      schema:
        type: string
        format: password
    filePath:
      name: path
      in: query
      required: true
      description: Chemin du fichier ou répertoire sur le serveur FTP
      schema:
        type: string

# Sécurité par défaut sur toute l'API (Authentification de l'utilisateur FlopBox)
security:
  - basicAuth: []

paths:
  /api/server/:
    post:
      summary: Ajouter un nouveau serveur FTP
      tags:
        - Serveurs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alias:
                  type: string
                  example: "MonServeur"
                host:
                  type: string
                  example: "ftp.example.com"
                port:
                  type: integer
                  example: 21
      responses:
        '201':
          description: Serveur créé avec succès

    get:
      summary: Récupérer la liste de tous les serveurs FTP enregistrés
      tags:
        - Serveurs
      responses:
        '200':
          description: Liste des serveurs

  /api/server/{ID}:
    get:
      summary: Récupérer la configuration d'un serveur spécifique
      tags:
        - Serveurs
      parameters:
        - $ref: '#/components/parameters/serverId'
      responses:
        '200':
          description: Configuration du serveur
        '404':
          description: Serveur inconnu

  /api/server/{ID}/directory:
    get:
      summary: Afficher l'arborescence d'un répertoire
      tags:
        - Répertoires
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      responses:
        '200':
          description: Contenu du répertoire
        '401':
          description: Erreur d'authentification FTP
        '403':
          description: Accès refusé
        '404':
          description: Serveur ou chemin inconnu

    post:
      summary: Créer un nouveau répertoire
      tags:
        - Répertoires
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      responses:
        '201':
          description: Répertoire créé
        '400':
          description: Le répertoire existe déjà ou erreur de création

    put:
      summary: Renommer un répertoire
      tags:
        - Répertoires
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  example: "nouveau_nom"
      responses:
        '200':
          description: Répertoire renommé avec succès
        '400':
          description: Impossible de renommer
        '404':
          description: Répertoire introuvable

    delete:
      summary: Supprimer un répertoire
      tags:
        - Répertoires
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      responses:
        '204':
          description: Répertoire supprimé avec succès
        '404':
          description: Répertoire introuvable

  /api/server/{ID}/file:
    get:
      summary: Télécharger un fichier depuis le FTP
      tags:
        - Fichiers
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      responses:
        '200':
          description: Fichier trouvé et renvoyé
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Fichier introuvable

    post:
      summary: Téléverser (upload) un nouveau fichier
      tags:
        - Fichiers
      parameters:
        - $ref: '#/components/parameters/serverId'
        - name: chemin_cible
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Fichier téléversé avec succès
        '400':
          description: Fichier invalide ou erreur d'upload

    put:
      summary: Renommer un fichier
      tags:
        - Fichiers
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  example: "nouveau_nom"
      responses:
        '200':
          description: Fichier renommé avec succès
        '404':
          description: Fichier introuvable

    delete:
      summary: Supprimer un fichier
      tags:
        - Fichiers
      parameters:
        - $ref: '#/components/parameters/serverId'
        - $ref: '#/components/parameters/filePath'
        - $ref: '#/components/parameters/ftpUsername'
        - $ref: '#/components/parameters/ftpPassword'
      responses:
        '204':
          description: Fichier supprimé avec succès
        '404':
          description: Fichier introuvable

  /api/server/{ID}/search:
    get:
      summary: Chercher un fichier/répertoire
      tags:
        - Recherche
      parameters:
        - $ref: '#/components/parameters/serverId'
        - name: query
          in: query
          required: true
          description: Mot-clé pour la recherche
          schema:
            type: string
      responses:
        '200':
          description: Recherche aboutie
        '400':
          description: Requête invalide